+++ src/steps/characterList.ts
@@ -6,16 +6,18 @@ import {
   homePage,
 } from "../toyhouse";
 import { Character, characterList } from "../util/db";
+import { progressIndicator } from "../util/progressIndicator";
 
 /**
+ * [Character List] step
  * Generate a character list. Expects browser to be on
- * the toyhouse homepage and logged in.
+ * the toyhouse homepage and logged in. Sets up the lightweight
+ * characterlist db entry for consumption by other processes.
  * @param page
  */
 export async function processCharacterList(page: Page) {
-  // onward to the character list
+  // navigate to the user's all characters page
   const profileLink = await page.evaluate(homePage.getNavLinkUrl, "Profile");
-
   await page.goto(`${profileLink}/characters/folder:all?page=1`, {
     waitUntil: "networkidle2",
   });
@@ -39,7 +41,9 @@ export async function processCharacterList(page: Page) {
     );
 
     logger.info(
-      `Found ${charactersOnPage.length} characters on page ${currentPage}/${pageCount}`
+      `${progressIndicator(currentPage, pageCount, 0)} Found ${
+        charactersOnPage.length
+      } characters on page ${currentPage}`
     );
 
     charactersOnPage.forEach((c) => {
@@ -54,6 +58,6 @@ export async function processCharacterList(page: Page) {
   }
 
   await characterList.set(characters);
-
+  logger.info(`Successfully indexed ${characters.length} total character(s)`);
   return characters;
 }